Revert Chinese changes because xxxx said so
Except for the clerical change of fixing the accelerators. I can't judge
which translation is better, but the changes are overall not consistent
with the rest of the translation, changing up stuff like punctuation or
spacing where multiple ways of doing it are apparently valid, so there's
no point in switching things partially over to a different system.