Implement OpenGL-based canvas
Significantly faster than the QGraphicsView-based canvas, especially in
the browser. Allows the user to pick which canvas renderer they want to
use in the preferences. This commit reworks a bunch of the
infrastructure to allow both interfaces simultaneously, the old variant
set to be phased out eventually. A software renderer via QPainter should
be implemented as well, but using the same interface as the GL canvas.