Implement QPainter-based software canvas renderer
Using mostly the same interface as the OpenGL canvas, with the split
into controller, view and renderer. The performance seems to be
virtually identical to the QGraphicsView canvas, so this mostly achieves
a reduction in code duplication between the old and new canvas view.

This also renames the "OpenGL" option in the settings to "Hardware".
That's clearer and more accurate, since ANGLE and WebGL are sort of but
not quite OpenGL.