Start new process when opening a new window
On Linux and Windows, since that's more resilient to crashes. If
replacement fails, it falls back to just spawning a window in the
current process. The behavior can be disabled by setting the
`DRAWPILE_SINGLE_PROCESS` environment variable to a non-empty, non-zero
value.

On Android and the browser, this isn't done because they can't start new
processes. On macOS it's not done because I think it would be
unexpected, since it's got that process-wide Mac menu, which is
presumably something that Mac users expect to have.

This adds several new command-line options to control what the
application does on startup.