Nice errors for unavailable drawpile-srv options
If the server has been compiled without libmicrohttpd, libsodium or
QtWebSockets and the user tries to pass an option related to those, the
server will now give a pretty error message telling them that they can't
use this option because it's not compiled in. The help output also now
informs the user that those options aren't available because stuff is
not compiled in.

Previously, those options were just not compiled in if they were
unavailable, meaning the server just told the user that they passed an
unknown option, which didn't give a good clue as to what was wrong.