Remove private listings and roomcode lookups
They are a vestigial feature that no longer makes sense and likely no
one is using them anymore in the presence of invite links. The only
thing they do is confuse users by claiming that their session is
"private", when it still may be listed publicly because of listserver
includes. This removes the feature both from the client and the server,
leaving only public listings.

Resolves #1224.