Fix macOS CI, build both Intel and Apple Silicon
The latest XCode broke building Qt5, so now macOS is on Qt6. This
unfortunately ups the required minimum macOS version to 11, but not
really much we can do about it. Should hopefully not be too bad, since
that is over 10 years old.

This rejigs the main workflow a bit. Instead of using a matrix and then
a bunch of includes and excludes, it now specifies every option we want
to build explicitly, since that's easier to understand and makes it
actually possible to remove entries temporarily if you don't care about
them on a branch.

We build with both the Intel-absed macos-13 and ARM-based macos-14
runners now. Since we now have a platform that has two different 64 bit
tarets, we now specify the full architecture in the run name and
artifact names.

Fixes #1300.