Fix contrast on depressed buttons and expand icons
In the Qt Fusion style. Depressed buttons in dark theme are now darker
so that they are actually distinct. Expand arrow buttons on menus and
toolbars use a proper icon instead of some hard-coded thing that's too
dark to see in any dark theme.

Also adds a matching patch for Qt6 so that it can also get these
improvements, along with the checkbox outline contrast increase.