Don't reparent start dialog children when hosting
Because it triggers a crash on Qt 6.6+ on Linux for some reason, it
doesn't like that the dialog gets reparented and then immediately
deleted I guess. We can avoid that crash just by not reparenting in that
case, since it's pointless to do so anyway if the dialog is just gonna
be deleted afterwards.

Fixes #1305