Browser checking and standalone messages
The web browser version now shows a message taken from update.json in
the startup dialog, if one exists. Known issues and the development
warning are no longer shown. Browsers are now checked for and if the
user is using a known incompatible browser, they get an appropraite
warning. The application will also now start immediately if there's no
update available, no troublesome browser is used and no notice is to be
shown.