Rewrite selections and transforms
Selections and transforms are now distinct concepts.

Selections are persistent on the canvas, acting similar to layers. They
participate in the undo history. They can be modified while on the
canvas. Layers can be selected by their bounds or turned into
selections. Pixel opacity is taken into account.

Transforms are temporary, but act as multipart operations that can be
undone and redone while in progress.

Implements #335, #1009, #1116, #1284 and #1303.