Improve tool click detection algorithm
Taking into account which device is used, so mice have a much smaller
detection distance now than tablets or touch. This also reduces the
distance for the latter significantly, since it was so large that it
caught false positives too much.

This also moves the mouse smoothing application into the tool
controller, since it now knows what device is being used and can decide
whether to apply the smoothing or not, instead of having the canvas
controller do it.