Add PROXY_STYLE build option
To enable a proxy for the Fusion style that fixes the bad contrast
issues in dark themes without patching Qt itself. This causes trouble on
some locales of Windows, so it can't replace the patches, but is
necessary for systems where using the unpatched system Qt is the only
sane option, like Flatpak or the AUR package.