Allow transforming multiple layers at once
When starting a transform, you can now pick which layers to transform by
checking and unchecking them in the tree. Transforming a group will
check everything inside of it automatically. Layers that are locked or
censored get a forbidden icon instead so that it's clear that they're
not part of the transform.

The accurate transform preview will transform each layer separately, up
to at most 16 layers due to how the preview stuff is architected with
some fixed-length buffers that get atomically swapped around. Beyond
that, the preview is forced into fast mode, which just previews the
merged image. That should hopefully be fine, since previewing a huge
amount of layers with full accuracy gets pretty unbearably slow.

Stamping with multiple layers selected is not currently possible. It
could be implemented, but it doesn't seem sufficiently useful to me to
be worth the extra effort.

Implements #1158.