Don't look tools not affected by layer or canvas
Because it doesn't make sense to lock the annotation or zoom or select
tool just because you have a group layer selected or something. Those
tools don't act on layers in the first place, so it doesn't make sense
to lock them in that case.

This also now allows finishing a multipart operation on a layer that's
locked, because the tool may not want to actually modify the active
layer. It also didn't handle the canvas being locked, so the check
wasn't ideal in the first place. If we want to check permissions here,
the tool should be doing it instead, since it knows better what it's
doing.