Discourage psudo-Chrome on iOS/iPadOS
Like every browser, Chrome on those devices is just Safari in a
trenchcoat. It doesn't handle the view size correctly somehow, causing
parts of the window to end up outside of the screen sometimes. Not worth
fixing, so now we just tell the user to use the real Safari instead.