Preview fill and magic wand before applying it
Because they are somewhat lenghty operations that are annoying to undo,
so now they give you a chance to undo locally instead. This also makes
using the fill tool without a size limit sensible.

The fill tool previews as a checkerboard pattern on the target layer,
the magic wand as an item in the canvas scene.

This also gives the magic wand tool a size limit slider, to make the
whole UI analogous to the fill tool.