Adhere to system double click parameters
In the click detector for tools. If the system has a more lenient
double-click interval or distance than we would normally use, we defer
to that instead. If ours is more lenient, we use that. That's likely to
be the case for detecting click distance for touch and tablet, since we
make that more lenient due to the inaccuracy of clicks on those devices,
whereas the system parameter usually assumes a more stationary device
like a mouse.

Implements #1335.