Write saved files to temporary storage on Android
Before copying them over the real, truncated file. This is not as ideal
as QSaveFile since it's not atomic, but better than just clobbering it
and potentially crashing along the way. KARchive also doesn't support
QIODevice::Truncate, so this way we can use a plain QFile instead and
not be left with garbage leftovers on Android that are devastating for
ZIP files that put their index at the end.