Revert hardware renderer to default if it crashes
On at least one Windows 7 system, the hardware renderer is known to
crash when initializing. This is apparently due to a null pointer
somewhere inside libGLES2, which is part of ANGLE, which we can't
exactly do much about without digging deep into that code.

So for now, we just detect the crash and revert to the default renderer
if it happens. This is accomplished by setting the renderer
implementation to the default upon initializing the hardware renderer,
the initializing it, and then restoring the value again. If there's a
crash in the middle, the value won't be restored and the next run will
not try to use the hardware renderer again.