Allow disabling network proxies
Both manually and automatically. There's now a setting in the network
preferences to toggle between the system-defined proxy and disabling it
altogether. When the system-defined proxy is enabled, Drawpile will
check if it's valid at all and automatically disable it if not, e.g.
because it's a proxy that can't actually tunnel connections.

Relates to #1360.