Add menu button to color docks title bar
For the color wheel, this moves the settings for shape, angle and
direction to here instead of being in the tools preferences. Color space
is in both places, since it affects more than just the wheel.

For the color sliders, this adds settings for the color space, which
sliders to show and whether to show the color input.

For the color palette, this moves the menu that used to be in the row
below.

Also changes the left-side spacing of the titlebar to 2 instead of 6
pixels, since that makes it match up better with stuff in the second
row. I think.