Attach brushes to slots
Instead of presets just being presets, they can now keep track of
changes like they do in most other drawing programs. When selecting a
brush, it is attached to the slot and shown next to the preview. Changes
will automatically be remembered until reverted, the changes can also be
written to a new brush or clobber the current one. The separate brush
properties dialog is gone, its features have been folded into the brush
settings dialog. Brushes can still be detached and there's a setting in
the tool preferences to disable automatic attachment.

Missing feature: you can't overwrite a different brush with the current
one. Don't think that's particularly relevant though, so not bothering
to implement it right now.