Allow assigning brush presets to shortcuts
And also rejigging a bunch of other stuff with regards to shortcuts and
the settings dialog. The input tab is now the tablet tab instead and
there's a button at the bottom for the tablet tester. Analogously,
there's a touch tester button on the touch page. Brush presets are their
own tab in the shortcuts settings and allow assigning the same key
sequence to multiple brushes, which it will toggle through in sequence.
Conflicts are handled across all types of shortcuts now and you can
filter for them. The mirror and flip actions have gotten extra
descriptive text to allow searching for them as "mirror" and "flip",
since otherwise you may think you can only do one of those. Searching
for the keybinds themselves is also possible now.

Implements #1367, #1368 and #1377. Relates to #1386.