Allow compiling without default update checks
Via -DDISABLE_UPDATE_CHECK_DEFAULT=ON. This will change the setting
default to be off instead of on by default, but the user can still
enable it or manually check if desired. To be used for F-Droid, since
that client will notify users of updates on its own anyway and considers
update checks by default an antifeature.