Don't require confirmation for fill and magic wand
The fill tool now previews fills locally without an overly obvious
indicator to that regard and making another click will implictly finish
that fill and start a new one. There's now buttons to apply and cancel
the fill and he click to confirm can be brought back again in a menu in
the top-left corner.

The previewed fill no longer changes as you modify parameters by
default, since we want to pretend it's a "normal" fill tool and users
won't expect that e.g. changing color will modify an existing fill. That
can also be reenabled in the menu.

The magic wand now always applies immediately, since selections are
local-only anyway and it doesn't have that many useful parameters to
change to begin with. That may change again in the future.