Improve client too old error messages
It now tells them more clearly that their version is outdated and what
to do to update. Setting a minimum protocol version of at least 4.24.0
will also now implicitly require a session lookup, since all clients
with that version are supposed to support it, failing to do a lookup
will show a pretty error that the client version is outdated.