Drag adjustment of fill and magic wand tolerance
Similar to how the magic wand tool works in GIMP, clicking and dragging
left and right will temporarily adjust the tolerance and show a preview
of the effect thereof.

Note that this is not how the fill tool works in GIMP, that just seems
to have bogus behavior there when you drag it.

Implements #1327.