Start host dialog rework
Redesigns the host tab in the start dialog entirely. It now comes with
multiple tabs, allowing the setting of way more session settings in
advance of actually hosting it, making it less annoying to set up
sessions. As an ancillary change the undo limit has been raised to 60.

Relates to #1220.