Handle lifting and replacing fingers in multitouch
By kinda rewriting that part entirely so that it doesn't reset the zoom
and rotation when you lift and replace a single finger in a multitouch
operation. It now resets the zoom and rotation when the finger
configuration changes and remembers the "anchor" points of touch points
as the current point when that change happens, making all zooms and
rotations relative to that.