Use PlaySound on Windows for notification sounds
Because on at least one Windows 7 system the Qt implementation causes a
crash, which is not acceptable for the application making a bloop noise.
The implentation is now split between Qt5, Qt6 and Windows, where the
latter uses PlaySound from Winmm. The function is loaded dynamically, in
case there's systems without that DLL or something, in which case we
just don't play back any sound. Since the function doesn't support
setting a volume, we load the WAV file into memory and manipulate its
samples to change their volume, which is a bit silly, but works.