Pre-rotate MyPaint tilt inputs
To get sensible tilt results when rotating the canvas. Since we already
have to mirror and flip them this way in absence of any other way to
input that information into libmypaint, will probably give consistent
results. It's not what MyPaint does, but it also doesn't give proper
results when rotating the canvas.

Relates to #1423.