Make message handling reentrant
Because the browser somehow manages to receive more messages on the main
thread while it'ss already handling messages. Probably some kind of
async code problem, where the WebSocket handler manages to butt in and
cause the message receival to be reentered. The message handler now
checks for that and sets a flag telling it that there's more to handle
instead of nesting itself and causing mayhem.