Allow toggling a local-only sketch mode on layers
This mode allows setting a local opacity and color tint on the layer. By
default, it tints a light blue at 75% opacity. The tint is based on
lightness, so it should generally Do What You Mean even when the sketch
has a white background. The layer menu at the top and the layer context
menu have toggles for this. The layer properties dialog has a toggle,
opacity slider and a tint picker with two convenience presets of no tint
and the default blue. The layer tree dock has a button next to the
opacity slider to toggle the mode, activating it also shows a button to
change the tint and switches the opacity slider to one for the sketch.
The last changed color tint and opacity are saved in the settings, the
next activation will use those prior values. Sketch opacity and tint are
saved to and loaded from ORA files in Drawpile-namespaced properties on
layers.

Implements #960.