Add File > Export Again to export without reprompt
Exporting now remembers the path previously used and then allows
exporting again to the same path without reprompting. Not available in
the browser, because it only allows downloading images, or on Android,
because file paths are such a mess there that we don't need to add
another thing that can go wrong there.

Implements #1298.